Pressure Control Unit Interface App Code
This code is installed and run in a personal computer to communicate
with arduino microcontroller. The code includes functions to set PID
gains and send a desired pressure waveform. The code is coded in C sharp
(i.e., C++ ) code and can be installed as a standard Windows software,
uninstalled in the Control Panel of Windows. Once the software is
installed, one can find the program Valve Control Interface and
Communication GUI in Start up> All programs. Through a computer, this
code talks to an arduino board, which is already downlowed for the
embeded code, which will be described in the section 7. The GUI enables
the user to interact graphically and control the controller unit.
Communication elements such as BAUD rate are easily accessible from the
GUI. A graph in GUI shows the current pressure as it is received. Access
to a PID controller parametes is allowed and the target pressure can be
either constant or varying. The communication is done over a Serial
port. Significant parts of this program are segments of open source code
found over the internet called ZedGraph
(http://sourceforge.net/projects/zedgraph/).